# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GNNcab - Multi-city ride dispatch and driver tracking platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: Get trips (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: driverId
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: cityId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Request a new trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /driver/profile:
    get:
      operationId: getDriverProfile
      tags: [drivers]
      summary: Get current driver profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverDetails"
    put:
      operationId: updateDriverProfile
      tags: [drivers]
      summary: Update current driver profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverDetails"
  /driver/trips:
    get:
      operationId: getDriverTrips
      tags: [trips]
      summary: Get trips for current driver
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Driver trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
  /customer/profile:
    get:
      operationId: getCustomerProfile
      tags: [users]
      summary: Get current customer profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
    put:
      operationId: updateCustomerProfile
      tags: [users]
      summary: Update current customer profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
  /customer/trips:
    get:
      operationId: getCustomerTrips
      tags: [trips]
      summary: Get trips for current customer
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Customer trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - message
...
    DriverDetails:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        vehicleMake:
          type: string
          nullable: true
        vehicleModel:
          type: string
          nullable: true
        vehicleYear:
          type: number
          nullable: true
        vehicleColor:
          type: string
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        status:
          type: string
          enum: [AVAILABLE, BUSY, OFFLINE]
        isVerified:
          type: boolean
        currentLat:
          type: number
          nullable: true
        currentLng:
          type: number
          nullable: true
        lastSeen:
          type: string
          format: date-time
          nullable: true
        cityId:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalTrips:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - email
        - firstName
        - lastName
        - status
        - isVerified
        - totalTrips
        - createdAt
        - updatedAt
...
    UpdateDriverProfileRequest:
      type: object
      properties:
        phone:
          type: string
        licenseNumber:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: number
        vehicleColor:
          type: string
        vehiclePlate:
          type: string
...
    Trip:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        driverId:
          type: string
          nullable: true
        cityId:
          type: string
          nullable: true
        pickupLat:
          type: number
        pickupLng:
          type: number
        pickupAddress:
          type: string
        dropoffLat:
          type: number
        dropoffLng:
          type: number
        dropoffAddress:
          type: string
        status:
          type: string
          enum: [REQUESTED, DRIVER_ASSIGNED, DRIVER_ACCEPTED, DRIVER_REJECTED, DRIVER_EN_ROUTE, DRIVER_ARRIVED, TRIP_STARTED, TRIP_COMPLETED, TRIP_CANCELLED, DRIVER_NO_RESPONSE, TIMEOUT]
        distanceKm:
          type: number
          nullable: true
        estimatedFare:
          type: number
          nullable: true
        finalFare:
          type: number
          nullable: true
        currency:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, online]
        cancellationReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
        assignedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - pickupLat
        - pickupLng
        - pickupAddress
        - dropoffLat
        - dropoffLng
        - dropoffAddress
        - status
        - currency
        - paymentMethod
        - requestedAt
        - createdAt
        - updatedAt
...
    CreateTripRequest:
      type: object
      properties:
        pickupLat:
          type: number
        pickupLng:
          type: number
        pickupAddress:
          type: string
        dropoffLat:
          type: number
        dropoffLng:
          type: number
        dropoffAddress:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, online]
        notes:
          type: string
        cityId:
          type: string
      required:
        - pickupLat
        - pickupLng
        - pickupAddress
        - dropoffLat
        - dropoffLng
        - dropoffAddress
        - paymentMethod
...
    TripsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...
    CustomerProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        preferredPaymentMethod:
          type: string
          nullable: true
        totalTrips:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - email
        - firstName
        - lastName
        - totalTrips
        - createdAt
        - updatedAt
    UpdateCustomerProfileRequest:
      type: object
      properties:
        phone:
          type: string
        preferredPaymentMethod:
          type: string
          enum: [cash, card, online]
...